Changing svn-make_cesm_tag to use the revision at which the tag was made, rather than trying to obtain the revision from which the tag was copied, in printing the changes since the last tag. Here are my comments about this:
Note: I used to get the revision from which the tag was copied
(using the -r switch to svn-latest_tag) rather than getting the
revision of the tag itself. That is more robust for catching
problems, like some trunk commits not being included in the last
tag. However, it was giving some "false positives" (saying that some
changes weren't included in the last tag when they really were),
which seemed to stem from people using a workflow that I didn't
expect. So now I am just doing the simpler thing of looking at
changes since the last tag was made - this assumes that the last tag
captures all changes that were made up to that point on the
respective trunk / branch.

Also, deleting a no-longer-relevant note from the documentation of
svn-latest_tag


git-svn-id: https://svn-user-sacks.cgd.ucar.edu/bin/trunk@566 81eb0212-1297-45d0-9317-8d7c8fb0914c