Reverting r566 for svn-make_cesm_tag: going back to the more robust method of determining the revision from which a tag was made.
Point of doing this is: sometimes I am making tags from a revision
that is not the latest trunk revision. Also, other people seem to be
shifting to making tags the "right" way, removing the motivation for
doing this in the other (simpler) way.

(Note: I did NOT revert r566 for svn-latest_tag)



git-svn-id: https://svn-user-sacks.cgd.ucar.edu/bin/trunk@591 81eb0212-1297-45d0-9317-8d7c8fb0914c